flask_app/requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
Flask==2.3.3
Flask-Session==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
Werkzeug==2.3.7
